+++ tests/regressions.test.js
@@ -283,7 +283,7 @@ describe('optional markup does not crash a component', () => {
         <li class="active"><img src="http://localhost/1.jpg"></li>
         <li><img src="http://localhost/2.jpg"></li>
       </ul></div>`;
-    const instance = Expressive.Slider.init(document.querySelector('.slider'));
+    const instance = Expressive.Slideshow.init(document.querySelector('.slider'));
 
     try {
       instance.set(1);